Migration Fixes #1

+++ database/migrations/2017_06_10_062632_add_foreign_keys_to_invoice_items_table.php
@@ -2,6 +2,7 @@
 
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
 
 class AddForeignKeysToInvoiceItemsTable extends Migration
 {
@@ -13,7 +14,12 @@ class AddForeignKeysToInvoiceItemsTable extends Migration
     public function up()
     {
         Schema::table('invoice_items', function (Blueprint $table) {
-            $table->foreign('invoice_id')->references('id')->on('invoices')->onUpdate('RESTRICT')->onDelete('RESTRICT');
+            $sm = Schema::getConnection()->getDoctrineSchemaManager();
+            $indexesFound = $sm->listTableIndexes('invoice_items');
+            if(!array_key_exists("invoice_items_invoice_id_foreign", $indexesFound)){
+                $table->foreign('invoice_id')->references('id')->on('invoices')->onUpdate('RESTRICT')->onDelete('RESTRICT');
+
+            }
         });
     }
